Update Naming Convention and Push to Production
Previously the files composing the Conceptualizer
were named with a scheme designed
to help me keep track of the checklist for the final project,
and were accordingly under the folder
"Checklist".
Rename the folder "Conceptualizer", rename
all files within it to be descriptive of their purpose,
and fix all imports.
Additionally, change descriptive text and router links
to reflect the naming "Conceptualizer" rather than
either "Checklist" or "The Inacurrate Conception".